+++ sparc/include/u.h
@@ -54,8 +54,8 @@ union FPdbleword
 {
 	double	x;
 	struct {	/* big endian */
-		ulong hi;
-		ulong lo;
+		uint32_t hi;
+		uint32_t lo;
 	};
 };
