+++ sparc/include/ureg.h
@@ -9,46 +9,46 @@
 
 struct Ureg
 {
-	ulong	r0;			/* unnecessary; just for symmetry */
+	uint32_t	r0;			/* unnecessary; just for symmetry */
 	union{
-		ulong	sp;		/* r1 */
-		ulong	usp;		/* r1 */
-		ulong	r1;
+		uint32_t	sp;		/* r1 */
+		uint32_t	usp;		/* r1 */
+		uint32_t	r1;
 	};
-	ulong	r2;
-	ulong	r3;
-	ulong	r4;
-	ulong	r5;
-	ulong	r6;
-	ulong	r7;
-	ulong	r8;
-	ulong	r9;
-	ulong	r10;
-	ulong	r11;
-	ulong	r12;
-	ulong	r13;
-	ulong	r14;
-	ulong	r15;
-	ulong	r16;
-	ulong	r17;
-	ulong	r18;
-	ulong	r19;
-	ulong	r20;
-	ulong	r21;
-	ulong	r22;
-	ulong	r23;
-	ulong	r24;
-	ulong	r25;
-	ulong	r26;
-	ulong	r27;
-	ulong	r28;
-	ulong	r29;
-	ulong	r30;
-	ulong	r31;
-	ulong	y;
-	ulong	tbr;
-	ulong	psr;
-	ulong	npc;
-	ulong	pc;
-	ulong	pad;	/* so structure is double word aligned */
+	uint32_t	r2;
+	uint32_t	r3;
+	uint32_t	r4;
+	uint32_t	r5;
+	uint32_t	r6;
+	uint32_t	r7;
+	uint32_t	r8;
+	uint32_t	r9;
+	uint32_t	r10;
+	uint32_t	r11;
+	uint32_t	r12;
+	uint32_t	r13;
+	uint32_t	r14;
+	uint32_t	r15;
+	uint32_t	r16;
+	uint32_t	r17;
+	uint32_t	r18;
+	uint32_t	r19;
+	uint32_t	r20;
+	uint32_t	r21;
+	uint32_t	r22;
+	uint32_t	r23;
+	uint32_t	r24;
+	uint32_t	r25;
+	uint32_t	r26;
+	uint32_t	r27;
+	uint32_t	r28;
+	uint32_t	r29;
+	uint32_t	r30;
+	uint32_t	r31;
+	uint32_t	y;
+	uint32_t	tbr;
+	uint32_t	psr;
+	uint32_t	npc;
+	uint32_t	pc;
+	uint32_t	pad;	/* so structure is double word aligned */
 };
