+++ sys/src/9/k10/etherif.h
@@ -53,7 +53,7 @@ struct Etherpkt
 
 extern Block* etheriq(Ether*, Block*, int);
 extern void addethercard(char*, int(*)(Ether*));
-extern ulong ethercrc(uchar*, int);
+extern uint32_t ethercrc(uchar*, int);
 extern int parseether(uchar*, char*);
 
 #define NEXT(x, l)	(((x)+1)%(l))
